requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.2
filelock==3.12.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
llvmlite==0.40.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
openai-whisper @ git+https://github.com/openai/whisper.git@e8622f9afc4eba139bf796c210f5c01081000472
outcome==1.2.0
packaging==23.1
PySocks==1.7.1
python-dotenv==1.0.0
regex==2023.8.8
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.12
tiktoken==0.3.3
torch==2.0.1
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
urllib3==2.0.4
webdriver-manager==4.0.0
wsproto==1.2.0
